Allow for multiple instantiations of mysql-router
Before this change it was not possible to run more than one
instantiation of mysql-router on a single unit. The Nova cells
deployments require nova-cloud-controller to be connected to two
different databases, therefore, requiring two instantiations of
mysql-router.

This change includes all the necessary steps to run multiple
instantiations of mysql-router on a single unit.

Review and merge charm-helpers PR:
https://github.com/juju/charm-helpers/pull/469

Closes-Bug: #1876188
Depends-On: Iafcc106fca44479e89b4b66a0a3988ffeee01f04
Change-Id: I76d1bfdccf9a551188c9f86c06ee0a22c07d47cf