Stable rebuild to fix dependency issue
The charm in the store has regressed due to a change in a upstream
dependency.

This issue was fixed by pinning said dependency in layer-basic [0].

0: https://github.com/juju-solutions/layer-basic/pull/179

Move pinning of stable deps to wheelhouse-override as the current
approach adds both versions and the one with the newest version
will be used.

Add explicit netaddr pin in wheelhouse-override as the order of
dependencies may override the one consumed from layer-basic.

Change-Id: I33d7f81319cd3515d4dac9a91b001281d3155bbf
Closes-Bug: #1884460