Add connect_timeout to MySQLHelper.connect call
If the local mysql router fails to connect (or work) with the clustered
MySQL8 instance, then the MySQLHelper.connect() call will hang
indefinitely.  Unfortunately, this happens during an update-status hook,
which is not visible (due to juju status not showing update-status hooks
executing).  Although this has no impact on the running cloud, it does
mean that nothing else happens on the machine (update-status, upgrades,
hooks, config-changed, etc.).  This allows the connect to fail and thus
for the charm to go into error so that the local mysql router instance
can be restarted.

Note that this depends on [1] landing in charm-helpers.

[1]: https://github.com/juju/charm-helpers/pull/550

Closes-bug: #1907250
Change-Id: I030d7caa8cc41aaf89ade25e5be196fb602d466c