Set client_ssl_mode
When vault certificates are being used and passed to the mysql-router
via the db-router relation use client_ssl_mode = PASSTHROUGH. This
maintains the < 8.0.23 behavior terminating TLS at the
mysql-innodb-cluster nodes.

If not using vault certificates from the db-router relation use
client_ssl_mode = PREFERRED, 8.0.23's new behavior. This TLS encrypts
with self-signed certificates. This covers TLS for the deployment
process.

If ssl_ca is unset on the db-router relation, unset it on the shared-db
relation to guarantee the requestor uses the correct communication
method.

Closes-Bug: #1914299
Change-Id: I1e66aca79df83b41072bb5df2cfb1708c8259cb4
Co-authored-by: Aurelien Lourot <aurelien.lourot@canonical.com>