Use force on subsequent attempts to add instance
If for any reason the first attempt to add an instance to cluster fails,
subsequent attempts would fail with:

Error: It appears that a router instance named <name> has been
previously configured in this host. If that instance no longer exists,
use the --force option to overwrite it.

This change avoids this error by adding the --force switch on subsequent
attempts.

Closes-Bug: #1919560

Change-Id: Icb0b10d6a5e3fe902c7ea56dbcb2f04d1129cd5f