Configure /run/mysql location for PID file
Ensure that a location in /run is used for the PID file for
the mysqlrouter instance - this ensures that in the event
of a power outage, the old PID file will not be present and
the daemon will correctly startup when the machine starts up.

Update systemd configuration to create /run/mysql and to use
the mysql user and group to execute the daemon.

Switch to 'forking' mode to ensure that the mysqlrouter process
gets tracked by systemd.

Closes-Bug: 1931095
Change-Id: Id3f7605ad1e9e4a6e4b0a1abe8000abd75b5b0fd
(cherry picked from commit 72aca90ea3551a95ecf0bf301fd626bfb8b25a97)