Ensure update config function can handle new headers
The linked bug occurred due to a new header for the config file being in
the upgraded charm, but not in the previous version of the charm.  As
the config file is read, updated, and then written, new headings in the
INI file would crash the charm code with a KeyError.

This patch just sets the header and parameter if it is missing.

Closes-Bug: #1927981
Change-Id: I18a6a4143ee0a1144eade5caa50611b802cba28a