Restart router if connections fail with 2003 code
At the moment if a connection through the router fails after a
configuration update the router is only restarted if the connection
error has a code of 2013 but often the error thrown is 2003 (see
*1) . This patch alters the charms
behaviour to also restart the router on a 2013 error.

While testing this patch it became apparent that a connection
attempt through the router immediatly after the router has been
restarted very often fail. So, the connection attempt has been
moved into its own method with its own tenacity retry logic.
A side effect of this is that the total possible wait time
has increased from 5 * 10 (outer tenacity loop) to 5 * 10 * 5
(outer tenacity loop and inner tenacity loop).

*1 https://dev.mysql.com/doc/mysql-errors/8.0/en/client-error-reference.html

Closes-Bug: #1973177
Change-Id: I9c2846bf4f21d2dcb1958bee4c9fa72dd4464b6c