Add docs key and point at Discourse
Add the 'docs' key and point it at a Discourse topic
previously populated with the charm's README contents.

When the new charm revision is released to the Charmhub,
this Discourse-based content will be displayed there. In
the absense of the this new key, the Charmhub's default
behaviour is to display the value of the charm's
'description' key.

Change-Id: I803a39bbddf129b0b8e2a882fcab5b7afedbbd68
(cherry picked from commit 1c77791a984f8865d93f722fc4273674224fb3ae)