Add cleanup of configuration files on `stop`
When this subordinate is removed from a unit it leaves its configuration
files behind. This can become an issue if the subordinate is added back
to the unit as pointed out in #2004088.

This change adds a cleanup step to the `stop` hook so that those
configuration files are removed on the removal of the subordinate charm.

Closes-Bug: #2004088
Closes-Bug: #2004089
Change-Id: Icfca352f29e089aeccd1434d33a573ca1682339a
Signed-off-by: Nicolas Bock <nicolas.bock@canonical.com>